fix: Prevent user from changing genre name to one that exists
When editing a genre, the application allowed a user to enter
the name of another existing genre. Since a unique constraint
exists in the database, this change would fail showing an
unfriendly SQLite error. Instead of allowing attempts to
commit such a change, we should instead catch this error
before committing and flash an informative error message to
the user.